Modernize cmake, remove unsupported and dead code
Get closer to a working Windows build

+++ src/Proto.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2018 The Open-Transactions developers
+// Copyright (c) 2019 The Open-Transactions developers
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -10,8 +10,8 @@ namespace opentxs::proto
 std::string ToString(const ProtobufType& input)
 {
     std::string output{};
-    output.resize(input.ByteSize());
-    input.SerializeToArray(output.data(), output.size());
+    output.resize(static_cast<std::size_t>(input.ByteSize()));
+    input.SerializeToArray(output.data(), static_cast<int>(output.size()));
 
     return output;
 }
